requirements.txt
openai
faiss-cpu
python-dotenv
numpy
fastapi
uvicorn
pytest
pydantic